Change the esquema of the project in backend
