+++ Backend/utils/utils.go
@@ -1,0 +1,13 @@
+package utils
+
+import (
+	"fmt"
+	"strings"
+)
+
+func ParsePrice(price string) float64 {
+	var value float64
+	price = strings.Replace(price, "$", "", -1)
+	fmt.Sscanf(price, "%f", &value)
+	return value
+}
